+++ src/Controller/VideosController.php
@@ -79,12 +79,13 @@ class VideosController extends AppController {
     }
 	
 	/**
-	 * List videos by a date
+	 * Lists videos by a day (year, month and day).
+     * It uses the `index` template.
 	 * @param int $year Year
 	 * @param int $month Month
 	 * @param int $day Day
 	 */
-	public function index_by_date($year, $month, $day) {		
+	public function index_by_day($year, $month, $day) {		
 		//Sets the cache name
 		$cache = sprintf('index_date_%s_limit_%s_page_%s', md5(serialize([$year, $month, $day])), $this->paginate['limit'], $this->request->query('page') ? $this->request->query('page') : 1);