+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2.4 branch
 ### 2.4.0
 * added action to list videos by month (year and month) and by year;
-* added "videos by month" widget;
+* added "videos by month" and "videos by year" widgets;
 * renamed `description` field as `text`;
 * `index_by_date` action renamed as `index_by_day`.